lumiai-backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# AI 服务
google-genai==1.0.0

# CORS 和安全
python-multipart==0.0.6

# 工具
python-dotenv==1.0.0
